+++ sfc-api/src/main/java/com/xiaotao/saltedfishcloud/service/collection/CollectionService.java
@@ -27,7 +27,7 @@ public interface CollectionService {
      * @param uid   调用者用户ID，用于验证权限
      * @param cid   文件收集ID
      */
-    void deleteCollection(int uid, long cid);
+    void deleteCollection(long uid, long cid);
 
     /**
      * 关闭一个收集任务，停止收集
@@ -35,7 +35,7 @@ public interface CollectionService {
      * @param cid   收集ID
      * @return  关闭后的收集任务信息
      */
-    CollectionInfo setState(int uid, Long cid, CollectionInfo.State state);
+    CollectionInfo setState(long uid, Long cid, CollectionInfo.State state);
 
     /**
      * 创建收集任务
@@ -43,7 +43,7 @@ public interface CollectionService {
      * @param info  收集任务信息
      * @return      收集任务ID
      */
-    CollectionInfoId createCollection(int uid, CollectionDTO info);
+    CollectionInfoId createCollection(long uid, CollectionDTO info);
 
     /**
      * 获取一个收集信息
@@ -69,5 +69,5 @@ public interface CollectionService {
      * @param submitFile    提交的文件信息
      */
     @Transactional(rollbackFor = Throwable.class)
-    void collectFile(CollectionInfoId cid, int uid, InputStream is, FileInfo fileInfo, SubmitFile submitFile, String ip) throws IOException;
+    void collectFile(CollectionInfoId cid, long uid, InputStream is, FileInfo fileInfo, SubmitFile submitFile, String ip) throws IOException;
 }
